Make menu items only appear once historic data uploaded and processed
This will prevent users from clicking into menu items that won't work before the processing is complete.

This makes use of polling (as websockets sounded complex to use). It's not ideal because this does mean requests are made every 15 seconds until the data is processed, but by keeping track of state and previous state for the relevant processing state variable, we can prevent the api calls after the processing step is complete.