Added classes and class creation for ease of performing calculations
Rather than working directly with the Django objects, I've created distinct classes for shifts, role types and rotas. While this does mean there's some quite odd code duplication (i.e. each model field has to be respecified as an attribute), this does allow for leveraging the powers of Python classes, including class methods and creating additional attributes that are useful for calculations or display. This also makes it easier to transfer work to and from notebooks used for prototyping.