Create README.md
# CalculateAvailableTime.ipynb

This notebook contains the beginnings of time calculations. 
It loads in some json data - exported directly from API requests, so it should be easy to port this code back into the app - and then creates objects from them to work with. 

# Recreating Excel Model.ipynb

This was used when manipulating the data set ready for Prophet forecasts.

Also contains code for assessing forecast quality. 

# record_level_from_excel.ipynb

This was used to generate a shareable record-format dataset from the aggregated data that was available in the existing shared Excel model. 

It keeps the original trends on an hourly level, with additional columns generated with data that has a random level of modification to provide some variety.